testing_project_quant_native_openspec/requirements.txt
akshare>=1.12.0
baostock>=0.8.8
tushare>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.18.0
pandas-ta>=0.3.14b
pytest>=7.4.0
PyYAML>=6.0
mplfinance>=0.12.0
matplotlib>=3.7.0
